RESOLVED - bug 194138: cleanup unnecesary casts and null checks https://bugs.eclipse.org/bugs/show_bug.cgi?id=194138

+++ org.eclipse.mylyn.trac.tests/src/org/eclipse/mylyn/trac/tests/TracXmlRpcTest.java
@@ -201,8 +201,8 @@ public class TracXmlRpcTest extends TestCase {
 		call("ticket.milestone.create", "foo", attributes);
 
 		Map<?, ?> values = (Map<?, ?>) call("ticket.milestone.get", "foo");
-		assertEquals(new Integer(due), (Integer) values.get("due"));
-		assertEquals(new Integer(completed), (Integer) values.get("completed"));
+		assertEquals(new Integer(due), values.get("due"));
+		assertEquals(new Integer(completed), values.get("completed"));
 
 		call("ticket.milestone.delete", "foo");
 	}
